Added FloatField and changed min-max of LongField from doubles to longs, kept step as double.

+++ common/src/main/java/net/mysticcreations/lib/config/specification/LongField.java
@@ -1,8 +1,8 @@
 package net.mysticcreations.lib.config.specification;
 
 public class LongField extends ConfigField<Long>{
-    double min;
-    double max;
+    long min;
+    long max;
     double step;
 
     public LongField(Long defaultValue) {
